Add SBC decoder contract tests and vector generation tool
- Implemented contract tests for SbcDecoder in sbc_decoder.ino, validating decoding against generated SBC vectors.
- Created a YAML configuration for the sbc_decoder sketch to support both host and ESP32 profiles.
- Developed a Python script (gen_sbc_vectors.py) to generate SBC test vectors from a vendored Broadcom encoder, producing a header file with encoded frames and PCM data.
- Added a test script (test_sbc_decoder.py) to verify the output of the SbcDecoder tests.

+++ src/external/sbc/encoder/include/sbc_types.h
@@ -1,3 +1,9 @@
+/* NOTICE: Modified by PCMFlowBluetooth (tools/sync_sbc.py).
+ * 32-bit integer typedefs changed from `long` to fixed-width types so the
+ * codec is correct on 64-bit hosts. No-op on 32-bit targets such as the
+ * ESP32. See ../../../LICENSE_sbc.md.
+ */
+#include <stdint.h>
 /******************************************************************************
  *
  *  Copyright (C) 1999-2012 Broadcom Corporation
@@ -34,7 +40,7 @@
 #include "stack/bt_types.h"
 
 typedef short SINT16;
-typedef long SINT32;
+typedef int32_t SINT32;
 
 #if (SBC_IPAQ_OPT == TRUE)
